requirements.txt.txt
# Core dependencies
python-dotenv==1.0.0
openai==1.12.0

# LangChain
langchain==0.1.9
langchain-openai==0.0.5
langchain-community==0.0.20

# LangGraph (для агентов)
langgraph==0.0.20

# Telegram
aiogram==3.4.1

# Optional: LangSmith для трейсинга
# langsmith==0.0.87